Foreign check update total amount issue fixed

+++ Standing_Order_Vat_App.Common/ViewModels/ForeignCheckVm.cs
@@ -15,7 +15,7 @@ namespace GbRegister.Core.ViewModel
         public long RecordId { get; set; }
         [Required]
         [Display(Name ="Check Number")]
-        public int? CheckNumber { get; set; } 
+        public long? CheckNumber { get; set; } 
         [Required]
         [Display(Name = "Payer Account Number")]
         public long? PayerAcctNumber { get; set; }
